fix: template docs
